+++ src/main/java/kitchenpos/order/repository/OrderRepository.java
@@ -1,0 +1,5 @@
+package kitchenpos.order.repository;
+
+public class OrderRepository {
+
+}
