+++ src/main/java/kitchenpos/ordertable/domain/OrderTable.java
@@ -1,6 +1,5 @@
 package kitchenpos.ordertable.domain;
 
-import java.util.List;
 import java.util.Objects;
 import javax.persistence.Embedded;
 import javax.persistence.Entity;
@@ -12,7 +11,6 @@ import javax.persistence.ManyToOne;
 import kitchenpos.common.constant.ErrorCode;
 import kitchenpos.common.domain.Empty;
 import kitchenpos.common.domain.NumberOfGuests;
-import kitchenpos.domain.Order;
 import kitchenpos.tablegroup.domain.TableGroup;
 
 @Entity
