+++ src/test/java/kitchenpos/acceptance/OrderRestAssured.java
@@ -3,7 +3,7 @@ package kitchenpos.acceptance;
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
-import kitchenpos.domain.Order;
+import kitchenpos.order.domain.Order;
 import org.springframework.http.MediaType;
 
 public class OrderRestAssured {
