+++ src/test/java/kitchenpos/ordertable/application/TableServiceTest.java
@@ -9,8 +9,7 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
 import kitchenpos.dao.OrderDao;
-import kitchenpos.domain.OrderStatus;
-import kitchenpos.ordertable.application.TableService;
+import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.ordertable.domain.OrderTable;
 import kitchenpos.ordertable.domain.OrderTables;
 import kitchenpos.ordertable.dto.OrderTableRequest;
